Don't override the size passed in widget_settings
If this causes any problems, we'll have to discuss and revisit. But
going forward, we should just make sure there are sensible defaults
and that they are overridden when specific params are provided.

The other thing we should probably do at some point, if possible, is
try to adjust to different aspect ratios...