One language code mapping to rule them all, and in the darkness bind them.
This commit is the first step in consolidating our many mappings of languages
codes and names into something sane.

First, it moves utils/language_codes.py into its own library so it can be called
from settings.py (there's some stuff in utils/__init__.py that needs to happen
after settings.py, so it couldn't stay in utils/).

Next, it adds the concept of "name" and "native" name to unilangs (the new name
for language_codes).

The name of a language is a unicode string of the English name for the language.
It's marked for gettext translation, and should be used with gettext to provide
an appropriate version of a language's name to a viewer.

The native name is a Unicode string that will never ever change -- it's what
speakers of language X call language X.

Finally it changes settings.py to use unilangs instead of rewriting the list.

More to come.