Add a workaround to allow review in multiple steps.
This hacky fix allows reviwers and approvers to review their own subtitles if
they somehow manage to get assigned to the task.

This will happen when they create a new version while in the process of
reviewing.  It's an ugly hack but needed for certain partner workflows, and we
don't have time to really do it right (for now).