Fixed the Metadata rename migrations.
On migration  0138 we had the models frozen to the field names *before* the migration ran, not after.
So I've left 138 in the correct state (this should work because this will only be used on migrating the next one), and created a migration 139 for making sure we have a unique together constraint on subtitle version and metadata key.