Use latest raven release
This is how it went.  I used 1.8.1 and it was horribly broken because of some
backwards incompatible changes in celery.  So, I downgraded our raven version
to 1.7.0 before all this celery stuff was introduced.  I opened an issue on
Github and it got resolved in like 2h.  They immediately cut a minor release.
This is 1.8.2 which this commit introduces.