Fix #933 (i hope)
can_review queries the video latest_version to see
if the user of it is the same as the reviewer. if we
don't save it before verifying and the reviewer is made
some changes when reviewing the last version's user will be
the reviewer.
that means: if someone does a post-publish edit, we won't be
able to set the assignee because he is both the reviewer **and**
the author. if we save before verifying (and the post-publish
editor is not the reviewer) we will be able to set the task's
assignee