Fix #1076
If the reviewer changes the version and sends back,
and the translator doesn't make a change to the
subtitles, the task would not get assigned to the
reviewer, since we don't allow people to review
their own subtitles.
For this case we should override that param and
let the user review it.