Pull out some admin fields into properties.
On the surface this looks like a bad idea.  Before this change, Django would do
a select_related() when viewing the SubtitleLanguage and SubtitleVersion admin
pages to get the video title and language title respectively.  Now it runs
a query once per row instead.

This sounds like a regression, but for some reason MySQL was performing a full
table scan when the select_related was used.  So instead of one massive painful
query we'll now have 21 tiny queries.

In a perfect world we'd figure out why MySQL is doing a full table scan, but
this world isn't perfect, these pages aren't user-facing, and this solution
works.