Fixing #1142
The auto-create function sets the subtitle_version
of the task for a subtitle_version of another 
language, we need to verify if the sv is the same
language as the task language.

Because of this, the form was finding a sv for the
task and was not creating the language for the video.