Turn teams.views.activity into a normal, non-generic view.
Generic views hide away the pagination machinery, which we're going to need for
my next commit to work around MySQL's bad decisions.

Also, they just suck in general.  Invariably they need to be torn out at some
point, and it's not much more code to write them the same way as all the other
views in the first place.

Also, this change adds a metric so we can time the performance of the page.