Fix #1193
If the user that makes the review edits the
subtitles, we create the approve one with the "wrong"
review_base_version (as in, it's not the translate
one but the review). That means if the approve task
is sent back we were assigning it to the reviewer
(he was the user from review_base_version).

Remove the code that was assigning using review_base_version
so we always go after the user that translated/subtitled
the view.