Remove the doorman app
It was created and used a while ago, but we don't currently use it for anything.

Since it was created, there are a number of third-party apps that have been
open-sourced that perform the same function (Waffle, Gargoyle, probably others).
If we need this functionality in the future we should just use one of those.