Switch to databased-backed cached sessions.
Previously we were using the default, which was uncached, database-backed
sessions.  This meant that we had to hit the DB for every request by a logged in
user.

Now the sessions will be stored in memcache and the DB.  Reads will come from
memcache if possible, or the DB if not.