Added support for <b> and <i> on caption display / uploads and downloads, for #1182.
A few notes about this:
a) For the front end I hacked a bunch of old regexes instead of a proper parser. The nice markdown parsers in js required quite a lot of work to be compiled through the closure compiler. As we plan to remove closure soon and we only need a very small subset of markdown for now, it seems ok (even though I feel dirty).
b) Currently only SRTs are properly parsed on input and output, I'll do the remaining formats in different commits