Fix #1274
Always use get_subtitle_version, since autocreate
sets the subtitle_version to a language that it's
not the task's language and that's never updated.

This affects unpublish/publish because after publish
if we create a new version (by adding a subtitle,
for example) it makes the task's subtitle_version
be outdated.