Yet another fix for #1168
For some reason that i still don't know why,
sometimes TeamVideo search index for an instance
gets stale after a delete and that is making the
code on permissions to fail because we can't find
the instance from the search index.

If that happens, always return false on the permissions.
Also, if we a stale index, remove it from solr.