Fix #1482
We need to return the date the user saw the page,
not the page we created the last announcement.
This way, when the user clicks on hide, we will
use the date the user saw the page on the cookie
and if the user reloads, use that on the query
that finds if there's a new announcement.