Fix SubtitleLanguage.version
First, renames the version_no parameter to be version_number for consistency,
and fixes the places that currently use it (there aren't many).

Also fixes two bugs:

Asking for version number 0 would previously get the latest version because
0 happens to be falsy.  Now asking for version 0 will get you version 0 (which
is None because version numbers start at 1).

Also, there was no filtering on video, so if you did:

    sl = SubtitleLanguage.objects.filter(video=video, language_code='en')
    sl.version(version_number=4)

You would get a SubtitleVersion that had the version number 4 and was in
English, but could be from any arbitrary video the database decided to order
first.

Also adds a bunch of unit tests for most of the common use cases.