Add permissions checks for URLs on team videos (#1491)
I wasn't sure exactly where to put the permissions checking code.  Most of
the current code is in app.teams.permissions, but that code only works for
TeamVideo objects, but we needed a function that checked for any Video
object.  Putting it in the Video class wasn't great either because importing
from teams.permissions into that module creates a circular dependency.

So I created a new module called videos.permissions.  Hopefully that makes
sense to do.