Show tasks for all projects if the user is looking at their own tasks (#1597)
Moved code to handle picking the default projects into its own function. I
figure if it gets a :( symbol, it's good to keep it contained to one place at
least.

If the tedtalks project does't exist, then log a warning and ignore it.

Moved code in team_tasks() below where we initialize the filters variable.
It seems useful to be able to use that.