Make the filter/sort bar in the videos list work like the tasks list (#1597)
This changes when we show the clear link.  The task list showed it when the
user specified filters, while the videos list showed it when we were using a
filter.

This mainly affects when the user first navigates to the page.  In that case,
they don't have any filters specified, but for TED, we default to filtering on
the tedtalks project.  In that case we don't want to show the clear link.

The other case is when the user specifies "all projects".  In that case we do
want to show the clear link.