Fix #1664
When the user accessed a video with only one language
that was incomplete while on Chrome, the dialog
would auto select the incomplete language even though
it was disabled.

Now, instead of showing the Translated from field
based on all languages, show it based only on the
enabled languages.
Also, block the subtitling on rpc.py if the language
has no versions.