Fixes for video titles (1611)
Removed the TeamVideo.title attribute, we weren't really using it anywhere and
it's been replaced by the title in the subtitle version.

Changed the Video.title attribute so that it gets set to the latest title for
the original language subtitles.  This is the value that we've been
displaying, now the data in the DB matches it.

Added a test for the latter.