Move some team functionality into the subtitle pipeline.
Several changes here.

First, the pipeline can now take two extra arguments:

* committer, which is the person actually adding the subtitles.  This will be
  used for permission checks and such later on, and is currently used to decide
  some workflow based things.
* complete, which is a boolean that will set up the
  SubtitleLanguage.subtitles_complete attribute correctly.  This is also needed
  when determining some workflow based stuff.

Both of these are currently optional.

Second, the pipeline now handles the moderation of subtitle versions and the
creation of tasks.  Kind of.  There's probably still more we need to add in
here, but at least this is a decent start.