Refactor Youtube OAuth flow.
Instead of only being able to add a Youtube account to a team, we can now add
them to both teams and users.  I changed the "state" value that we send to
Youtube to be a JSON-string that contains the necessary information to
distinguish between teams and users.

We also make sure that the user who started the action is the one who is
finishing it.