Added a nose plugin to make tests run faster/more isolated.
Right now the plugin replaces a couple of tasks with mock objects.  It
should be pretty easy to extend it to other things.

Added a cache backend for test purposes.  Since we're only running in
1 thread, we can simply used a dict without any locking.

Fixed a couple tests to work with the new system.  In particular, if
you depend on the search index being updated, you need to call
test_utils.update_search_index.run_original().