Fix bleach assertion errors.
When we go to strip tags from the text of a caption, sometimes the bleach
library will raise an AssertionError.  This error has been recorded about 66k
times in production.

It happens when an empty html tag is stripped.

    >>> html = "<b></b>"
    >>> value = bleach.clean(html, strip=True, tags=[], attributes=[])
    >>> value
    u""

In this case, "value" will be an empty string but the logger will record an
AssertionError.  The error is very cryptic but harmless.  I have included a
unittest to prove that the correct value is returned by this call.  The
so-called fix to this problem is to no longer send the logged data to
sentry/stdout.  It tells us nothing useful.

The call to "bleach.clean()" only happens once in our codebase (apart from
unittests) so it should be easy to find when reviewing this.

https://sentry.amara.org/amaraproduction/group/936/