Don't die on VideoTypeError
When a Youtube video can't be found for some reason, we raise a VideoTypeError.
However, the get_or_create_for_url method on Video doesn't catch it and the
whole thing blows up.

The most common error is here:

https://sentry.amara.org/amaraproduction/group/283/

You can reproduce the error by visiting:

http://www.universalsubtitles.org/api/1.0/subtitles/?video_url=http://www.youtube.com/watch?v=VnkJM-UnK10

The fix is to catch the exception and return None from the method above.  If we
can't find the video on the provider's website, chances are we aren't going to
able to save it---so we bail early.