Fixed task assigning
Sometimes request.GET.assignee is empty and that
was not allowing our ajax call that assigns tasks (!)
to actually assign it (but it didn't fail) and
that was making review/approve tasks to fail
to finish.

Also added a check to verify that assignee is indeed
valid before saving the task.