1874: Handle blank language codes in the widget RPC
In the future these blank codes will be going away, but for now there are a few
hundred SubtitleLanguage's in our DB that have a language code of "".
Traditionally this meant "the original language of the video", so we'll just
make the name "Original" for now.

Done as part of Sentry Duty:
https://sentry.amara.org/amaraproduction/group/262/