1769: Fixes issue
Instead of hiding the whole control for anyone
that doesn't have access to add videos (can_edit_videos
is the same as can_add_videos), hide/display each
control depending on the right permission.

So now, users can see the tasks control as long as
they are members of the team (the same permission
as the top tab).