Implemented a resolve_ownership method on TPA
This method receives a VideoUrl and tries to find
the TPA account related to that VideoUrl.

We need this because we have a ton of video urls
pointing to youtube full names (instead of the
unique usernames) and since full names are not
unique, there could be two TPAs and that can
make the code blow up.