DMR: tear out unused methods on old models
These models should not be used any more, but we've decided to keep them around
just in case we want to look at the in the shell or admin.  As a compromise
we're going to remove all the extra functionality in their methods because it
should never be used anyway, plus it will make grepping around for things much
easier because you won't have all the extra false positives.