2105: First pass of flash fallback player for h264 videos on Firefox.
Currently  Firefox will not play <video> elements encoded with h264. This is changing and Firefox will support h264 on certain situations. Until that happens, we're using this fallback flash player in order to allow Firefox users to view those videos.

Talking to Popcorn folks, the old way of creating players (Popcorn.player()) is deprecated, and 'Wrappers' are the future. So this is the path we've taken.

I basically took their Vimeo player (actually the Vimeo media element) and ported to the flowplayer. This is why the code does actually much more than what we need ). However, if we ever what to improve on this player (for example with our own controls) most things should be in place.