2005: refactor unpublish view to handle deletion
Also pretty much rewrites the way tasks are created after unpublishing of any
kind, because now that we can delete it's just easier to look at the resulting
state and figure out what to do instead of looking at the action we performed.

This is a really tricky thing and there are probably bugs, so be careful.