Fixes the test_moderated_notifies_only_when_published
I think that test was only passing because it contained 2 logic errors
that were canceling each other out.

The first error was that it didn't account for the
team-task-approved-published.html email.

The second error was that before 824690c we were creating a new
SubtitleVerison when videos got approved, but calling
video_changed_tasks() using the old SubtitleVerison PK.  After the
change, we were now calling video_changed_tasks() with the correct key
and sending a new email.