Replaced the unpublish button
We decided to ditch the unpublish button and all the complexities that
go with it.  Instead, we now have a button to delete an entire
subtitle language

This deletes all SubtitleVersions in that language, as well as any
dependent languages.  This is the nuclear option for unpublishing
subs, and should only be used in extreme cases (like a language whose
only versions are spam).