2161: open correct editor when resuming forked translate tasks
The issue before was that for translate tasks, we would always open
the start dialog and let the user choose which language we were
translating from.  This is obviously not correct once the language has
been forked and we are resuming the task.

Changed it so that when we resume tasks, we skip the start dialog, and
go straight to the editor dialog.