2322: don't force a language for translation tasks
This fails when there are multiple languages to choose from since we
can't know which one to choose.  Here are a couple use cases for that:

  - Team transcribes into a language, starts several translations,
    then deletes the primary language and forks the translations.
    Team admin goes back and re-creates the transcribe task.

  - User creates a video with several languages.  The video gets moved
    to a team.  Team admin creates a transcribe task.

If there is one language, then the idea was to get the transcription
started as quickly as possible.  But even here, I'm not really sure
it's a good idea, what if the task asignee doesn't speak that
language?