2355: add "needs editing" tag for subtitles that have been sent back
Updated the language list tests a bit for tasks.  We should have be
using the primary language if we are testing subtitle tasks.

Updated Task.complete() so that it returns a task for when things get
sent back.  Before it was returning None.  It seems like that method
should return a new task if one was created.