2358: Don't crash when uploading translations from non-public languages
Changed it so that the form raises a ValidationError in this case,
instead of crashing when self.from_sv is None.

Changed it so the only valid choices for source translations are
public languages.  This matches the choices when you start the task in
the editor, so I think it's correct.