Refactor code to track the working subtitles
Changed it so that SubtitleController simply sets this on the root
scope.  We already had 2 different controllers that tracked this and
the timeline is going to be a third.  It seems cleaner to do this in
SubtitleController and remove the duplication.

Added an isWorkingSubtitles() method to SubtitleController.  I think
this makes the code more obvious than checking the isEditable attribute.