Added a class to deal with lists of subtitles.
Give SubtitleListController a class to handle subtitles list instead
of just storing an array.  This should make it easier to define
operations on the subtitles lists, like the ones I'm about to need for
the timeline.

Made dfxp.js put its variables inside a module.  I think this is
better than before, and it was also the easiest way for me to get
access to jquery using $.