Move timeline directive code inside the link function.
I'm not sure what I was thinking before, but this code makes more
sense inside the function.  Otherwise if we have multiple timeline
elements for some reason, the variables will be shared between them.
Also it's just more convenient to have access to scope and the
container element all the time.