Made changes to the timeline update the subtitle data
Made a signal for drag actions on the timeline widget.

Changed parser.startTime()/parser.endTime() to input integers instead
of strings when we want to update times.

Update the subtitle data for the timeline when changes are made.

Fixed the timeline <a> tags so that clicking on them doesn't cause the
page to reload.

Made a few improvements to the data layer:
  - Make it so that we keep the same SubtitleList object around.  This
    makes it so that the subtitle id attributes change when subtitle
    updates are made and we refetch the list from the DFXP parser.
  - Added SubtitleList methdod to update the times.
  - Don't call digest() in the work-done handler, the caller should be
    responsible for doing that once all the changes are made.