Refactored how angular code interacts with the subtitle list.
Instead of using the AmaraDFXPParser, make all access go through
SubtitleList, which then uses AmaraDFXPParser to read/write the DFXP
DOM.  This simplifies a bunch of things.

Since we are now doing all updates through SubtitleList, we don't need
to constantly reload/refetch the subtitles from the parser.

Made SubtitleList simply read all subtitles on load and store them in
an array instead of storing the jquery queryset and parsing the nodes
as needed.  This is simpler, and it also works with subtitles that
are not stored in the order of their timings.

Changed SubtitleItem to have it's start/end times in milliseconds.
This means we can use subtitle.startTime where we used to use
parser.startTime(subtitle)

Store content as both HTML and markdown in SubtitleItem.

Removed subtitle-ready event.  We were only catching it in the
video overlay and we weren't really using it.  subtitle-selected is
the event that we really wanted to catch.

Removed link to the workspacetools.js file which has been deleted.

Removed the timeline-drag event and have the timeline directive make
the change directly.  I had the idea that directives shouldn't be
making changes to the model, but that's not really true.