Adjust timeline display at the extremes
If we are at the start/end of playback, position the timeline so
that the current playback position is in the middle.  Before we would
try to adjust things so that we didn't have empty space at the
left/right side.

Hide scrollbars on the page.  I'm not sure why, but things seemed to
be overflowing by 1px, which made the scrollbar appear, and that
screwed with the calculations to set the timeline width.