Implemented adding/removing subtitles
Added a bunch of code to SubtitleList to handle inserting/removing
subtitles.

Fixed code to track syncCount.  Made firstSubtitleAfter() pay
attension to that, otherwise it would fail when trying to do the
binary search if unsynced subs were in the list.

Positioned the add subtitle button at the top of the subtitle.  Before
it was positioned below the subtitle, which didn't make sense.

Removed key handling code for the tab button.  I don't think that's
how we want to handle tab and it's not in the spec.  If we need to
re-implement it, it shouldn't be hard.