Lots of changes/fixes for editing subtitles
Moved code to track events from the SubtitleList directive to the
SubtitleList controller.  I think this is a more natural place for it.
Also changed it so we use angular event handling rather than jquery
event handling.  Mixing the two was resulting in confusion when the
user clicked on the add/remove subtitles buttons.  We would handle the
add/remove click event, then also handle the click event that starts
editing mode.

Use jquery to show/hide the textbox rather than angular.  This makes
it so we can immediately give it focus, rather than relying on a
timeout.

Tried to generally simplify/cleanup the code that handles editing
subtitles.

Remove focusIndex code.  I'm not sure exactly what it was supposed to
do, but it didn't seem to have any effect.