Fixes for syncing subs with the up/down arrows.
Implemented changing start/end times for the synced subs.

Refactored the way we handle them.  Now we always store which
subtitles will be updated when the user asks to sync the start/end
time.  This will make it easier to implement the helper boxes on the
sides of the subtitles list.

Renamed firstSubtitleAfter() to indexOfFirstSubtitleAfter().  The new
name is pretty verbose, but it better describes what the method does.