SubtitleList cleanup
Most of this is in preparation of supporting editing a new language
with no existing data.

Don't try to load the initial language data if versionNumber = "".

Don't store versionNumber in SubtitleListController.  We don't use it,
and we won't have a version number if we're editing a new language.

Removed weird timeout code.  We don't need it if we call
SubtitleListFinder.register() early on in the subtitleList directive.

Handle no reference languages a little better.

Don't store version in the LanguageSelectionController.  We don't
actually use it, and we won't have it if there isn't a reference
language to use.