2441: Keep same height when subtitles are edited
Before we would be displaying a textarea that was exactly the right
height, but then there would be whitespace below it which would add
extra padding at the bottom.  This was ugly by itself, but it also
caused the remove button to change location depending on if the
textbox was visible or not, which caused issues with clicking it.