Refactored the SubtitleItem code.
Instead of a single SubtitleItem class, we now distinguish between
subtitles that are currently stored in the DFXP and subtitles that are
in the process of being edited.

This is a cleaner version of the old fake subtitle code that the
timeline used and we can use it to update the trayinfo as we change
things.