Made the subtitle template use the new subtitle-repeat directive
Made a new controller that handles all the subtitle items.
WorkingSubtitleItemsController handles a bunch of what used to be
mixed between SubtitleListController, SubtitleListItemController, and
mix of the old SubtitleListController and the subtitleListItem
directive.  Unlike the old SubtitleListItemController, we only create
1 controller/scope for the entire list of subtitles.

Moved code that sets attributes for the working subtitles to the
workingSubtitles directive (renamed from workingSubtitlesWrapper).

Added service that sets the selection range and made the
subtitleRepeatDirective use it.

Added subtitleRepeatDirective method that gets the LI element for a
given subtitle item.