Simplified language handling
Replaced languageName and languageIsRTL with a language class.  It's
simpler to have 1 variable than multiples, and this also fits with the
angular system that the scope holds references to the model instead of
is the modal itself (aka always have a dot in your variables).

Instead of having a isRTL attribute, just have a dir attribute.  This
lets us set the direction in the templates instead of having to have
code in the directives/controllers.