Another refactor/simplification
Instead of having other classes call setLanguage() on the subtitles
list scope, call that inside getLanguages() or initEmptyLanguage().

Instead of having the list selector fetch the subtitles, just call
getSubtitles().  This removes a some duplicate code.  The service
caches the subtitle data, so we don't need that functionality in the
language selector.