2458: implement progress bar support.
Removed "-edge" from the image names.  I think the names are a bit
more consistent now.

Added handling of mouseleave for the volume bar since I realized
we should have it when coding the progress bar.

Things look a bit ugly now, but that's just because I need to re-do
the slices.