Started work to move app-wide data out of the working subtitles scope.
It seems weird that we import this scope into other scopes.  If we
need a model from it, then that model should be in the application scope.

Moved allowsSyncing, currentEdit, canAddAndRemove to the root scope.  Renamed
"allowsSyncing" to "canSync" to make it match canAddAndRemove.

For info on the working language, it never changes.  So we can just
setup things using django templating and keep it out of angular
altogether.

Moved the scroll-to-subtitle signal to the root scope.  Made it only
handled by the working subtitles list.