Finished moving things out of workingSubtitles scope
Again, the reason for this is that it seems weird for scopes to import
other scopes.  If data is app-wide, then it should be in the app
scope.  Also, we were using the SubtitlesListController for both the
working subtitles and the reference subtitles, but much of it's
functionality was only for the working subs.

So here's the new system:
  - There's a SubtitleVersion model which stores things like the
    subtitles list, the language/version number, title/description,
    etc.
  - We create 2 instances of that in the app scope, referenceSubtitles
    and workingSubtitles.
  - SubtitleListFinder is no longer.  Where we used to use that, we
    just use either referenceSubtitles, or workingSubtitles.
  - SubtitleListController is gone as well.  We don't need it at all
    anymore for the reference subs.  Merged the rest of the code with
    the WorkingSubtitlesItems controller into a new class called
    WorkingSubtitlesController.
  - The app scope handles saveSubtitles()
  - The subtitles-editor directive initializes the working subtitles
    instead of the subtitle-list directive.
  - We store the video id and version number of the working subtitles
    in the app scope.

Minor fix for the subtitles-repeat directive so that we can use
angular experessions to find the subtitle list instead just an
attribute name.