Refactored system to get data from python to the new editor
The old system was a fairly confusing mix of adding data to the HTML
attributes and the editorData javascript value.  It also had the
drawback that some of the data got set in directives, so it wasn't
available when the controller code ran.

The new system is that we put everything in the editorData variable.
Then we present that to angular with the EditorData service.  We store
the videoId in the application scope.  If components want the language
code or the version number, they can use the workingSubtitles model.