Implemented workflow and refactored controllers.app module
Implemented modal class to handle the workflow and a controller to handle
the workflow progression area.

Decided to refactor the app controllor module since I wanted to test
the new code and there wasn't really a good way to do that.  Also the
file was getting pretty large and needed some better organization.

Started to get a more serious with the unittests.  Added some a module
with some very basic mocking functionality.

I think this new system is probably how we should organize our other
controller modules.  Controllers aren't placed in the global scope,
but in a module.controller() call.  Modal classes are placed in a
value.  This makes everything easy to test.

Added signal to save the current version.

Removed SubtitleStorage from the WorkingSubtitleController.  There
isn't a dependency anymore.