Last details on the Workflow
Hide the timeline on startup and show it when we move to the sync
stage.

Have enter create a new subtitle in typing mode.

Only show sync helpers when the timeline is shown.

Make the keybord shortcuts depend on if the timeline is shown, not the
current workflow state.  This way users can manually hide/show the
timeline to get the behavior they want.

Changed CurrentEdit API to not need a LI element.  The controllers
shouldn't be forced to pass that in, and we can easily calculate it
from the directives.