Code reorganization
Moved it so that instead of {controllers, directives}/subtitles.js we
have video/{controllers, directives}.js.  If we only had one module,
then that gets put in the base directory.

This seems like a better way to group things, since it's groups things
into components instead of by what they are.  This is how we organize
the django modules after all.

Combined controllers/app.js and editor-app.js into a single app
module.

Made a workflow module and moved some of the code from app.js into it.

Moved SubtitleList class out of dfxp and into an angular module.

Remoted the root variables, we put things in angular modules instead
of the global namespace now.