2518: Allow JS compilation to skip using closure
Made it so that the default is not to use closure, if use_closure is
set, then we do.  Removed ignore_closure parameter, since use_closure
does the same thing.

If we don't use closure, then we just combine the files and run it
through jsmin.

Removed bundle_type from the param list for the compile_*_bundle
methods.  We know the type in those methods, so it's just redundant.

Removed unused USE_BUNDLED_MEDIA setting.  We need to use
COMPRESS_MEDIA instead.