Make SubtitleStorage.getLanguages() return Language objects
Before it was returning the raw data from the server.

Added versions attribute to Language.

Updated underscore version.  Stopped calling noConflict() in the
subtitles controllers file.  That made it inaccessible everywhere
else.  Use non-minified version of underscore for the subtitle editor,
we'll minify it ourselves when we build the media bundle.