2498: Initial work to get metadata in the editor
Add metadata for the current working subs in the editor_data variable.

Refactored code to get subtitles.  Made it use
format=json&sub_format=dfxp.  This way we can get things like
title/description/metadata from the response.

Changed the metadat h3's so we uppercase them with CSS.  This will
make things better once we translate the metadata type, since we don't
know if the translation will be uppercase or not.