2576: allow saving tasks without a version.
I'm not sure how our tasks got in this state, but it's definitely
normal for TED.  We have the get_subtitle_version() method to handle
this case and we've been using it as much as we can.

This is an ugly fix, but hopefully it holds up until we can ditch
tasks entirely and move to the new collab model.