2587: Fix resolve_ownership method
resolve_ownership should handle VideoURL objects having bad data in
account_username, but it was failing because the try..catch had bad
indentation.  While I fixed that, I also cleaned up the method a bit
and also made it so that it fixes the data in VideoURL if it can.

Replaced videos.tasks._add_amara_description_credit_to_youtube_vurl
with a mock function.  There's no reason that we need to run this task
in the tests.