Metadata format change.
Instead of a list of 2-tuples, just use a dict for metadata.  Dicts
are much nicer to use.  The reason to not do this was to preserve the
order that metadata that got added, but that seems very hypothetical
at this point.  Right now the only user of this feature is TED and
they only have 1 metadata field.

The only times the order will matter is when you're viewing it on a
page, and inside the editor.  For either case, I don't see it really
making much difference when order we display the metadata in.