Refactored Video.title_display()
title_display() was pretty weird.  It would truncate the title to 35 chars by
default (although it added "..." at the end, so really 38 chars), but most
uses of the function actually didn't want the truncation.  And many pieces of
code assumed that it wasn't truncated and performed their own trunctation.

Changed it so Video.title_display() returns the full title.
Added SubtitleLanguage.title_display() and SubtitleVersion.title_display()

Removed Video.get_title_display(), which was redundant and would
return the same value as video.title_display(truncate=False).

Refactored the code to calculate titles for the video templates
(video-view.html and subtitle-view.html).  Before we had several
(fairly complicated) pieces of code to calculate the title.
Refactored things so the 1 place we calculate it in is the
videos.views module.  This reduces duplicate code, makes things
easier to test, and easier to translate.  This changes our title META
tags very slightly.