Removed code that updates the search index based on widget views
I'm not 100% sure, but I believe that this was causing issues on the
site.  I would observe it running then quickly afterwards we would
have several different database transactions keeping locks on the DB
for long periods of time (> 10 minutes).

I think this has to do with the way that the code was structured, we
would try to index 200 videos at a time which takes quite a while.
Also since we would split things up into chunks, then send them all to
celery at once so we would have many workers updating the index at the
same time.

We don't really need this code because we re-index the videos once a
day in the search.tasks.update_search_index task.  In the long term,
I'm hoping to replace that task as well and restore the functionality
that I'm removing here: videos that get page views get their index
updated more often.