Speedup SubtitleVersion admin page (#788)
This code should be the best way to select subtitle versions from the
admin page.  Originally, we were doing a straight select of the
SubtitleVersion objects, which was not great because for each list row
we would end up doing extra queries to get the video/language.  Then I
changed it to call select_related(), but this ended up blowing things
up because it was joining to too many tables and the query was insane.
I believe it copied every row from the subtitles_subtitleversion table
into a temporary table (although maybe it was the videos_video table,
not 100% sure).

Anyways, this time around we are now calling select_related() for
exactly the videos and subtitle language table, which should make
things work great.